Move DateToString helper into lib/ (higher swimlane)
