+++ packages/liveblocks-core/src/types/CommentReaction.ts
@@ -1,4 +1,4 @@
-import type { DateToString } from "./DateToString";
+import type { DateToString } from "../lib/DateToString";
 
 export type CommentUserReaction = {
   emoji: string;
