+++ packages/liveblocks-core/src/types/InboxNotificationData.ts
@@ -1,4 +1,4 @@
-import type { DateToString } from "./DateToString";
+import type { DateToString } from "../lib/DateToString";
 
 export type InboxNotificationThreadData = {
   kind: "thread";
