+++ packages/liveblocks-core/src/types/InboxNotificationDeleteInfo.ts
@@ -1,4 +1,4 @@
-import type { DateToString } from "./DateToString";
+import type { DateToString } from "../lib/DateToString";
 
 export type InboxNotificationDeleteInfo = {
   type: "deletedInboxNotification";
